already_trained_model/already_trained_test/requirements.txt
# Stable Diffusion with Hugging Face diffusers
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
